tools/requirements-python.txt
pre-commit
black
pycparser
watchdog
python-Levenshtein
cxxfilt
mapfile-parser==2.1.4
tabulate
requests
graphviz
splat64==0.24.6
crunch64
spimdisasm>=1.26.0
rabbitizer==1.12.0
n64img==0.3.3
pygfxd
pillow
